requirements.txt
streamlit>=1.31.0
python-docx>=0.8.11
reportlab>=4.0.7
openai>=1.55.3
anthropic>=0.35.0
requests>=2.31.0